Minor changes.
Create a helper method in the work session model to know whether it's closed.

Seperate the action of session ending and adding a note. There is no real value to it but it brings some problems with validation in serializators.

Don't mark the note field as readonly but check changes during validation. Thus we can return an error instead of just refuse to change the note silently.

We don't actually need any custom input to create a new session, so there is no need to parse the request body.

103212490