Refactor tests.
Test serializer behavior with a serializer test. A mock request can be provided so there is no reason to lift up the test to the API level.

I still don't know about looking for a request object in the serializer context. It feels wrong but it's the most convenient way so far.